Fixed Commands with DepFiles not cleaning up
UpdateDirtyState was checking that ALL inputs were missing to trigger a clean up (including the dep file intputs) but that's not correct, the dep file inputs don't matter. As long as all the static inputs are missing, it should clean up.

Also fixed the Cooking Log that wasn't showing "(Cleanup)" for cleanup commands.