Fixed gMatchPath (again!)
`str.find(pattern)` was incorrect (because it'll include the wildcars in the search), but so was `str.find(pattern[0])` (because the next char in the pattern might be found before where the pattern starts matching again, and then we'll incorrectly decide it's a failure).

Example: Matching `medium_house/texture_albedo.png` with `*_albedo.*`: after the first `*`, `pattern[0]` is `_` and is found at `_house` (then `_albedo` doesn't match and it's a fail).

We need to instead search for the sub-string that is until the next wildcard. In the example that's `_albedo.`.

Also added a debug window to see which Input Filter passes/fails.