Fixed ERROR_JOURNAL_ENTRY_DELETED error
Previous change that tried to still use the cached state when it's too old introduced a bug: mNextUSN was still overwritten with the value read from the cache, but that value might be so low that it's not in the journal anymore. Since we rescan everything, we don't need to touch mNextUSN in this case.