Fixed edge case were outputs not being written was not detected
If a command cooked for another reason than the inputs changing (eg. rule version changed, or clicked the Cook Now button), the LastCookUSN wouldn't change (since it was only based on inputs) and the command was always considered not dirty as a result. Now the USN of the previous outputs is also taken into account, so we detect if an output is not written.

Also removed the code "estimating" a LastCookUSN in UpdateDirtyState() when it's the initial scan. It's not needed anymore since commands are always dirty because of a rule version mismatch in that case. In hindsights using the outputs USN there was also a bit weird, using the max input USN as usual would have been fine.